+++ service-connectors/examples/target-stream/main.tf
@@ -6,6 +6,7 @@ module "target_stream_connector" {
   providers = {
     oci = oci
     oci.home = oci.home
+    oci.secondary_region = oci.secondary_region
   }
   tenancy_ocid = var.tenancy_ocid
   service_connectors_configuration = var.service_connectors_configuration