+++ service-connectors/examples/target-stream/providers.tf
@@ -20,6 +20,16 @@ provider "oci" {
   private_key_password = var.private_key_password
 }
 
+provider "oci" {
+  alias                = "secondary_region"
+  region               = var.secondary_region
+  tenancy_ocid         = var.tenancy_ocid
+  user_ocid            = var.user_ocid
+  fingerprint          = var.fingerprint
+  private_key_path     = var.private_key_path
+  private_key_password = var.private_key_password
+}
+
 terraform {
   required_version = ">= 1.3.0"
   required_providers {